exported pathExists
